backend/pyproject.toml
[tool.poetry]
name = "orka-backend"
version = "0.1.0"
description = "Orka — AI Intelligence Platform Backend"
authors = ["Orka Team"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.111.0"
uvicorn = {extras = ["standard"], version = "^0.29.0"}
sqlalchemy = "^2.0.30"
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic = {extras = ["email"], version = "^2.7.1"}
pydantic-settings = "^2.2.1"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
httpx = "^0.27.0"
celery = {extras = ["redis"], version = "^5.4.0"}
redis = "^5.0.4"
cryptography = "^42.0.7"
scikit-learn = "^1.5.0"
pandas = "^2.2.2"
python-multipart = "^0.0.9"
tenacity = "^8.3.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.2.0"
pytest-asyncio = "^0.23.6"
httpx = "^0.27.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
